[TM-2008] move constants
