+++ libs/database/src/lib/constants/framework-tf.ts
@@ -1,0 +1,2 @@
+export const FRAMEWORK_TF_TYPES = ["terrafund", "terrafund-landscapes", "enterprises"];
+export type FrameworkType = (typeof FRAMEWORK_TF_TYPES)[number];
