+++ libs/database/src/lib/constants/landscapes.ts
@@ -1,0 +1,2 @@
+export const LANDSCAPE_TYPES = ["Ghana Cocoa Belt", "Greater Rift Valley of Kenya", "Lake Kivu & Rusizi River Basin"];
+export type LandscapeType = (typeof LANDSCAPE_TYPES)[number];
