+++ libs/database/src/lib/constants/organisations.ts
@@ -1,0 +1,2 @@
+export const ORGANISATION_TYPES = ["for-profit-organization", "non-profit-organization"];
+export type OrganisationType = (typeof ORGANISATION_TYPES)[number];
